LayoutTests/performance-api/performance-observer-basic-expected.txt

Basic Behavior test for PerformanceObserver APIs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Inside PerformanceObserver callback
PASS argumentsLength === 2 is true
PASS list instanceof PerformanceObserverEntryList is true
PASS obs instanceof PerformanceObserver is true
PASS obs === observer is true
FAIL thisObject instanceof PerformanceObserver should be true. Was false.
FAIL thisObject === observer should be true. Was false.

PerformanceObserverEntryList APIs
PASS list.getEntries() instanceof Array is true
PASS list.getEntries().length === 2 is true
PASS list.getEntries()[0] instanceof PerformanceEntry is true
PASS list.getEntries()[0].name is "mark3"
PASS list.getEntries()[1].name is "mark4"
PASS list.getEntries()[0].startTime <= list.getEntries()[1].startTime is true
PASS list.getEntriesByType() threw exception TypeError: Not enough arguments.
PASS list.getEntriesByType("not-real").length === 0 is true
PASS list.getEntriesByType("mark").length === 2 is true
PASS list.getEntriesByType("mark")[0] instanceof PerformanceEntry is true
PASS list.getEntriesByType("mark")[0].name is "mark3"
PASS list.getEntriesByType("mark")[1].name is "mark4"
PASS list.getEntriesByName() threw exception TypeError: Not enough arguments.
PASS list.getEntriesByName("not-real").length === 0 is true
PASS list.getEntriesByName("mark1").length === 0 is true
PASS list.getEntriesByName("mark3").length === 1 is true
PASS list.getEntriesByName("mark3")[0] instanceof PerformanceEntry is true
PASS list.getEntriesByName("mark3")[0].name is "mark3"
PASS list.getEntriesByName("mark4").length === 1 is true
PASS list.getEntriesByName("mark4")[0] instanceof PerformanceEntry is true
PASS list.getEntriesByName("mark4")[0].name is "mark4"
PASS list.getEntriesByName() threw exception TypeError: Not enough arguments.
PASS list.getEntriesByName("not-real").length === 0 is true
PASS list.getEntriesByName("mark1").length === 0 is true
PASS list.getEntriesByName("mark3").length === 1 is true
PASS list.getEntriesByName("mark3")[0] instanceof PerformanceEntry is true
PASS list.getEntriesByName("mark3")[0].name is "mark3"
PASS list.getEntriesByName("mark4").length === 1 is true
PASS list.getEntriesByName("mark4")[0] instanceof PerformanceEntry is true
PASS list.getEntriesByName("mark4")[0].name is "mark4"
PASS list.getEntriesByName("mark3", "not-real").length === 0 is true
PASS list.getEntriesByName("mark3", "mark").length === 1 is true
PASS list.getEntriesByName(null, "mark").length === 0 is true
PASS list.getEntriesByName(undefined, "mark").length === 0 is true
PASS successfullyParsed is true

TEST COMPLETE
